+++ corpus-reader/src/components/react-reader/settings/LayoutSettings.tsx
